refactor: modularize sports hub
